TaxGroup

This function identifies the membership of a node to a certain higher-ranking taxon.

Usage

TaxGroup(g, T_table, g_tax, p_tax)

Arguments

g

An igraph object.

T_table

Taxa table as data.frame

g_tax

The taxonomic level at which the nodes will be colored.

p_tax

The taxonomic level at which the network was built.

Value

Returns an igraph object with nodes colored based on their taxonomic rank.

Examples

# The input can be a net pre-processed by net_inference
# This example uses david_stool data set

library (seqtime)
data ("tc_dstoolA")
g <- net_inference (tc_dstoolA, "aracne")
#> Loading required package: SpiecEasi
#> 
#> Attaching package: ‘SpiecEasi’
#> The following object is masked from ‘package:igraph’:
#> 
#>     make_graph
#> Loading required package: minet
#> 
#> Attaching package: ‘minet’
#> The following object is masked from ‘package:SpiecEasi’:
#> 
#>     clr
#> Warning: the standard deviation is zero

TaxGroup (g, T_table = david_stool_lineages, g_tax = "Order", p_tax = "Genus")
#> Error in `[.data.frame`(T_table, , g_tax): undefined columns selected