Abundances matrix function

This function colors nodes by abundance at a taxonomic level.

Usage

abs_mat (abs.list, g.list, n)

Arguments

abs.list

Abundance list, the same lists you used to generate the network.

g.list

List of igraph objects (multilayer network).

n

Vertex number. Notice that if n>20 the nodes won't show too big.

Value

Returns a matrix, each column is a layer, and the values of the rows corresponds to the size of each node.

Examples

library(seqtime)
data ("mlnet_dstoolAB")

# Create abundance list
lst <- list(david_stoolA_otus, david_stoolB_otus)

abs_mat (abs.list = lst, g.list = mlnet_dstoolAB, n=10)
#> Warning: data length [519] is not a sub-multiple or multiple of the number of rows [152]
#>             [,1]      [,2]
#>   [1,] 1.1294446 1.2745045
#>   [2,] 1.1745465 1.2280428
#>   [3,] 1.1663971 1.1711030
#>   [4,] 0.7290293 1.2651451
#>   [5,] 1.1668141 1.2521416
#>   [6,] 1.2111339 1.2129483
#>   [7,] 1.2268080 1.2245394
#>   [8,] 1.2485116 1.2383055
#>   [9,] 1.2011243 1.2719148
#>  [10,] 1.2005015 1.1979592
#>  [11,] 1.1868717 1.1611603
#>  [12,] 0.1386294 1.2362059
#>  [13,] 1.1749412 1.2413514
#>  [14,] 1.2327498 1.1236829
#>  [15,] 1.1748503 1.2473692
#>  [16,] 1.2110420 1.2402424
#>  [17,] 1.2199642 1.2241189
#>  [18,] 1.1922150 1.1959187
#>  [19,] 1.1967803 1.1783319
#>  [20,] 1.0108956 1.2021483
#>  [21,] 1.1963855 1.2039002
#>  [22,] 1.0993933 1.2384675
#>  [23,] 1.2178439 1.2014422
#>  [24,] 1.2507802 1.2049723
#>  [25,] 1.2668463 1.0791255
#>  [26,] 1.1969667 1.0140021
#>  [27,] 1.2321573 1.0534227
#>  [28,] 1.2436085 1.0867005
#>  [29,] 1.2518965 1.0989960
#>  [30,] 1.3123690 1.1107345
#>  [31,] 1.2788260 1.1075211
#>  [32,] 1.2469805 1.1692301
#>  [33,] 1.2577440 1.0908064
#>  [34,] 1.2787542 1.0661322
#>  [35,] 0.1098612 1.1193959
#>  [36,] 1.1950309 1.1023388
#>  [37,] 1.2265478 1.1273703
#>  [38,] 1.2615062 1.1036598
#>  [39,] 1.2572170 1.1027556
#>  [40,] 1.1601550 1.1365909
#>  [41,] 1.0972276 0.3663562
#>  [42,] 1.1818047 1.0907734
#>  [43,] 1.2284047 1.1072434
#>  [44,] 1.2200718 1.1176711
#>  [45,] 1.2836544 1.1000081
#>  [46,] 1.2410641 1.1440452
#>  [47,] 1.2565511 1.1675350
#>  [48,] 1.2256737 1.1380662
#>  [49,] 1.2452944 1.1485914
#>  [50,] 1.2285078 1.1873575
#>  [51,] 1.2666335 1.1802014
#>  [52,] 1.2500990 1.1479638
#>  [53,] 1.2036328 1.1575008
#>  [54,] 1.2730233 1.1889531
#>  [55,] 1.3091273 1.2137708
#>  [56,] 1.2770147 1.1843739
#>  [57,] 1.2828288 1.0283361
#>  [58,] 1.2775050 1.1056919
#>  [59,] 1.2721617 1.0973443
#>  [60,] 1.2260343 1.0651549
#>  [61,] 0.2708050 1.1178138
#>  [62,] 1.2727633 1.1177187
#>  [63,] 1.1999227 1.1550192
#>  [64,] 1.1991257 1.1437879
#>  [65,] 1.1923478 1.0072133
#>  [66,] 1.2135838 0.9755104
#>  [67,] 0.9888019 0.8862059
#>  [68,] 1.2009121 1.0307452
#>  [69,] 1.2430691 0.9583558
#>  [70,] 1.2135500 1.0254778
#>  [71,] 1.1640280 0.9070503
#>  [72,] 0.9932804 1.0422817
#>  [73,] 1.0521830 1.1398782
#>  [74,] 1.1886866 1.1835893
#>  [75,] 1.0220522 1.1788593
#>  [76,] 1.0442551 1.1418692
#>  [77,] 1.0417269 1.1648208
#>  [78,] 1.1816439 1.1350289
#>  [79,] 1.1661570 1.1509269
#>  [80,] 1.1244078 1.1774020
#>  [81,] 1.1139744 0.7451822
#>  [82,] 1.1550414 0.9469237
#>  [83,] 0.7030857 1.0404566
#>  [84,] 1.1773103 0.7849324
#>  [85,] 1.1751336 1.0418285
#>  [86,] 1.2041481 0.8482809
#>  [87,] 1.1623876 1.0848482
#>  [88,] 1.1876380 1.0916996
#>  [89,] 1.1713005 1.0928364
#>  [90,] 1.2131602 1.0491052
#>  [91,] 1.2231643 1.0781910
#>  [92,] 1.1853020 1.0588325
#>  [93,] 1.1792033 0.9890402
#>  [94,] 1.1373893 1.0569392
#>  [95,] 1.1727440 1.0523876
#>  [96,] 1.1948053 1.0686864
#>  [97,] 1.2186650 1.1137403
#>  [98,] 1.2631347 1.1126675
#>  [99,] 1.2404325 1.1350112
#> [100,] 1.0507612 1.1328450
#> [101,] 1.0235270 1.0928525
#> [102,] 1.1792306 1.0840992
#> [103,] 1.1821439 1.1527548
#> [104,] 1.2168983 1.1389107
#> [105,] 1.2220321 1.1396695
#> [106,] 1.2411704 1.1551049
#> [107,] 1.1923942 1.1497873
#> [108,] 1.2412818 1.1831852
#> [109,] 0.1386294 1.1837072
#> [110,] 1.2050132 1.1889456
#> [111,] 1.2380286 1.2272544
#> [112,] 1.2526267 0.8862200
#> [113,] 1.2399371 0.9030137
#> [114,] 1.2410865 1.0383349
#> [115,] 1.1323712 1.0585978
#> [116,] 1.1682314 1.0642397
#> [117,] 1.1649430 1.0131021
#> [118,] 1.1794142 1.0679780
#> [119,] 1.2192597 1.1050795
#> [120,] 1.2009833 1.1197050
#> [121,] 1.2094677 1.1804975
#> [122,] 1.2605674 1.1689455
#> [123,] 1.1160911 1.2246173
#> [124,] 1.1447277 1.2300332
#> [125,] 1.1538993 1.2346382
#> [126,] 1.0561241 0.1098612
#> [127,] 0.9066355 1.2110965
#> [128,] 1.2236321 1.2673072
#> [129,] 1.1854442 1.2005076
#> [130,] 1.2468364 1.1957092
#> [131,] 1.3144113 1.2216063
#> [132,] 1.2656454 1.2581443
#> [133,] 1.2884474 1.2890124
#> [134,] 1.2638809 1.2803832
#> [135,] 1.3121471 1.1974538
#> [136,] 0.1098612 1.2320276
#> [137,] 1.3020158 1.2061018
#> [138,] 1.3331041 1.2305479
#> [139,] 1.2265760 1.2593581
#> [140,] 1.2223731 1.2370688
#> [141,] 1.2515427 1.2656731
#> [142,] 1.2217191 1.2990545
#> [143,] 1.2586705 1.1776589
#> [144,] 1.2056075 1.1841610
#> [145,] 1.2771540 1.1405797
#> [146,] 1.2313235 1.1275049
#> [147,] 1.2184415 0.2079442
#> [148,] 1.2729318 1.1885544
#> [149,] 1.1930345 1.2065643
#> [150,] 1.2436347 1.1898739
#> [151,] 1.2771693 1.1504661
#> [152,] 1.2483564 1.2131123