The function ctr_df generates a data.frame containing the degree of nodes for each layer.

The function ctr_df generates a data.frame containing the degree of nodes for each layer.

Usage

ctr_df (g.list, layer_names)

Arguments

g.list

List of igraph objects (multilayer network), preprocessed by ml_TaxGroup.

layer_names

Vector with the name of each layer.

Value

Returns a data.frame with centrality value (if preprocessed with ctr_ml or ctr_g: degree, betweenness or closeness), p_tax and g_tax of each node (the function takes this info from ml_TaxGroup).

Examples

#Load the example net
data ("mlTaxGroup_dstool")

# Create a vector with layers' names
v <- c("sample A", "sample B")

ctr_df (mlTaxGroup_dstool, layer_names = v)
#>     sample A sample B                                         p_tax
#> 1          1        0                                Nitrospiraceae
#> 2          0        0                                          none
#> 3          0        0                                 Rikenellaceae
#> 4          0        0                            Porphyromonadaceae
#> 5          0        0                                Bacteroidaceae
#> 6          0        0                                Prevotellaceae
#> 7          0        0                             Flavobacteriaceae
#> 8          0        4                             Acidobacteriaceae
#> 9          0        0                               Spirochaetaceae
#> 10         0        0                             Hyphomicrobiaceae
#> 11         0        0                             Bradyrhizobiaceae
#> 12         0        0                             Xanthobacteraceae
#> 13         0        0                           Methylobacteriaceae
#> 14         0        0                                  Rhizobiaceae
#> 15         0        0                              Caulobacteraceae
#> 16         0        0                              Rhodobacteraceae
#> 17         0        0                             Sphingomonadaceae
#> 18         0        0                                 Neisseriaceae
#> 19         0        0                                Rhodocyclaceae
#> 20         0        0                                Alcaligenaceae
#> 21         0        0                                Comamonadaceae
#> 22         0        0                              Burkholderiaceae
#> 23         0        0                              Oxalobacteraceae
#> 24         0        0                            Cardiobacteriaceae
#> 25         0        3                            Oceanospirillaceae
#> 26         0        0                                 Moraxellaceae
#> 27         0        0                              Pseudomonadaceae
#> 28         0        0                        Pseudoalteromonadaceae
#> 29         0        0                                Shewanellaceae
#> 30         0        0                                  Vibrionaceae
#> 31         0        0                               Pasteurellaceae
#> 32         0        0                            Enterobacteriaceae
#> 33         0        0                           Desulfovibrionaceae
#> 34         0        0                            Campylobacteraceae
#> 35         0        0                              Fusobacteriaceae
#> 36         0        0                             Coriobacteriaceae
#> 37         0        0                           Geodermatophilaceae
#> 38         0        0                                   Frankiaceae
#> 39         0        0                             Microbacteriaceae
#> 40         0        0                             Cellulomonadaceae
#> 41         0        0                            Intrasporangiaceae
#> 42         0        0                                Micrococcaceae
#> 43         0        0                              Dermabacteraceae
#> 44         0        0                                Dermacoccaceae
#> 45         0        0                              Actinomycetaceae
#> 46         0        0                            Bifidobacteriaceae
#> 47         0        0                          Propionibacteriaceae
#> 48         0        0                              Mycobacteriaceae
#> 49         0        0                                  Gordoniaceae
#> 50         0        0                            Corynebacteriaceae
#> 51         0        0                                  Nocardiaceae
#> 52         0        0                               Veillonellaceae
#> 53         0        0                                Peptococcaceae
#> 54         0        0                               Lachnospiraceae
#> 55         0        0         ClostridialesFamilyXIII.IncertaeSedis
#> 56         0        0                                Clostridiaceae
#> 57         0        0                         Peptostreptococcaceae
#> 58         0        0           ClostridialesFamilyXI.IncertaeSedis
#> 59         0        0                              Paenibacillaceae
#> 60         0        0                                   Bacillaceae
#> 61         0        0                                Planococcaceae
#> 62         0        0                                   Gemellaceae
#> 63         0        0                                 Aerococcaceae
#> 64         0        0                             Carnobacteriaceae
#> 65         0        0                              Leuconostocaceae
#> 66         0        0                              Lactobacillaceae
#> 67         0        0                               Enterococcaceae
#> 68         0        0                              Streptococcaceae
#> 69         0        0                           Erysipelotrichaceae
#> 70         0        0                               Ruminococcaceae
#> 71         0        0                             Staphylococcaceae
#> 72         0        0                                  Coxiellaceae
#> 73         0        0                               Nocardioidaceae
#> 74         0        0                                  Brucellaceae
#> 75         0        0                              Flexibacteraceae
#> 76         0        0                             Turicibacteraceae
#> 77         1        0                              Tsukamurellaceae
#> 78         0        0                            Phyllobacteriaceae
#> 79         0        0                             Brevibacteriaceae
#> 80         0        0                                Aeromonadaceae
#> 81         0        4                               Sinobacteraceae
#> 82         0        0                              Beijerinckiaceae
#> 83         0        0                              Catabacteriaceae
#> 84         0        0                                   Dietziaceae
#> 85         0        0                            Pseudonocardiaceae
#> 86         0        0                                           F16
#> 87         0        0                                Halomonadaceae
#> 88         0        0                              Xanthomonadaceae
#> 89         0        0                           Sphingobacteriaceae
#> 90         0        0                              Alteromonadaceae
#> 91         0        0                       Dethiosulfovibrionaceae
#> 92         0        0                                    Thermaceae
#> 93         0        3                            Bacteriovoracaceae
#> 94         0        0                              Acetobacteraceae
#> 95         0        0                                   Nostocaceae
#> 96         0        0                                 Colwelliaceae
#> 97         0        1                             Planctomycetaceae
#> 98         0        0                            Hydrogenophilaceae
#> 99         0        0                               Nocardiopsaceae
#> 100        0        0                             Rhodospirillaceae
#> 101        0        3                              Methylophilaceae
#> 102        0        0                         Promicromonosporaceae
#> 103        0        0                                          OM60
#> 104        0        0                                Cryomorphaceae
#> 105        0        3                              Methylocystaceae
#> 106        0        0                            Bdellovibrionaceae
#> 107        0        4                             Streptomycetaceae
#> 108        1        0                               Hyphomonadaceae
#> 109        0        0                               Solibacteraceae
#> 110        0        0                              Rubrobacteraceae
#> 111        0        0                                 Chromatiaceae
#> 112        0        0                                Bogoriellaceae
#> 113        0        0                                Williamsiaceae
#> 114        0        0                                Legionellaceae
#> 115        1        0                               Gallionellaceae
#> 116        1        0                              Armatimonadaceae
#> 117        0        0                                   Opitutaceae
#> 118        1        0                           Alicyclobacillaceae
#> 119        0        0                          Solirubrobacteraceae
#> 120        0        4                            Spartobacteriaceae
#> 121        0        0                             Aurantimonadaceae
#> 122        0        0                             Gemmatimonadaceae
#> 123        0        4                            Micromonosporaceae
#> 124        0        0                          Saccharospirillaceae
#> 125        0        0                                 Xenococcaceae
#> 126        0        0                           Verrucomicrobiaceae
#> 127        0        0                              Dermatophilaceae
#> 128        0        0                                  Listeriaceae
#> 129        0        0                             Psychromonadaceae
#> 130        0        0                                   Gemmataceae
#> 131        0        0                          Syntrophobacteraceae
#> 132        0        0                                 Haliangiaceae
#> 133        0        0                                Idiomarinaceae
#> 134        0        0                            Dehalobacteriaceae
#> 135        0        1                                       AKIW874
#> 136        1        0                                        ACK-M1
#> 137        0        0                                Eubacteriaceae
#> 138        0        0                                Deinococcaceae
#> 139        0        0                                Leptospiraceae
#> 140        1        0 ThermoanaerobacteralesFamilyIII.IncertaeSedis
#> 141        0        0                             Nitrosomonadaceae
#> 142        0        0                            Erythrobacteraceae
#> 143        0        0                                Thermicanaceae
#> 144        0        0                                Victivallaceae
#> 145        0        0                                           A29
#> 146        0        0                               Rhodothermaceae
#> 147        0        0                             Helicobacteraceae
#> 148        0        0                             Exiguobacteraceae
#> 149        0        0                             Patulibacteraceae
#> 150        0        0                           Methanobacteriaceae
#> 151        0        0                             Sanguibacteraceae
#> 152        0        0                                 Phormidiaceae
#>                         g_tax
#> 1                 Nitrospirae
#> 2                        BRC1
#> 3               Bacteroidetes
#> 4               Bacteroidetes
#> 5               Bacteroidetes
#> 6               Bacteroidetes
#> 7               Bacteroidetes
#> 8      Acidobacteria (phylum)
#> 9       Spirochaetes (phylum)
#> 10             Proteobacteria
#> 11             Proteobacteria
#> 12             Proteobacteria
#> 13             Proteobacteria
#> 14             Proteobacteria
#> 15             Proteobacteria
#> 16             Proteobacteria
#> 17             Proteobacteria
#> 18             Proteobacteria
#> 19             Proteobacteria
#> 20             Proteobacteria
#> 21             Proteobacteria
#> 22             Proteobacteria
#> 23             Proteobacteria
#> 24             Proteobacteria
#> 25             Proteobacteria
#> 26             Proteobacteria
#> 27             Proteobacteria
#> 28             Proteobacteria
#> 29             Proteobacteria
#> 30             Proteobacteria
#> 31             Proteobacteria
#> 32             Proteobacteria
#> 33             Proteobacteria
#> 34             Proteobacteria
#> 35      Fusobacteria (phylum)
#> 36    Actinobacteria (phylum)
#> 37    Actinobacteria (phylum)
#> 38    Actinobacteria (phylum)
#> 39    Actinobacteria (phylum)
#> 40    Actinobacteria (phylum)
#> 41    Actinobacteria (phylum)
#> 42    Actinobacteria (phylum)
#> 43    Actinobacteria (phylum)
#> 44    Actinobacteria (phylum)
#> 45    Actinobacteria (phylum)
#> 46    Actinobacteria (phylum)
#> 47    Actinobacteria (phylum)
#> 48    Actinobacteria (phylum)
#> 49    Actinobacteria (phylum)
#> 50    Actinobacteria (phylum)
#> 51    Actinobacteria (phylum)
#> 52                 Firmicutes
#> 53                 Firmicutes
#> 54                 Firmicutes
#> 55                 Firmicutes
#> 56                 Firmicutes
#> 57                 Firmicutes
#> 58                 Firmicutes
#> 59                 Firmicutes
#> 60                 Firmicutes
#> 61                 Firmicutes
#> 62                 Firmicutes
#> 63                 Firmicutes
#> 64                 Firmicutes
#> 65                 Firmicutes
#> 66                 Firmicutes
#> 67                 Firmicutes
#> 68                 Firmicutes
#> 69                Tenericutes
#> 70                 Firmicutes
#> 71                 Firmicutes
#> 72             Proteobacteria
#> 73    Actinobacteria (phylum)
#> 74             Proteobacteria
#> 75              Bacteroidetes
#> 76                 Firmicutes
#> 77    Actinobacteria (phylum)
#> 78             Proteobacteria
#> 79    Actinobacteria (phylum)
#> 80             Proteobacteria
#> 81             Proteobacteria
#> 82             Proteobacteria
#> 83                 Firmicutes
#> 84    Actinobacteria (phylum)
#> 85    Actinobacteria (phylum)
#> 86                        TM7
#> 87             Proteobacteria
#> 88             Proteobacteria
#> 89              Bacteroidetes
#> 90             Proteobacteria
#> 91              Synergistetes
#> 92                     Thermi
#> 93             Proteobacteria
#> 94             Proteobacteria
#> 95              Cyanobacteria
#> 96             Proteobacteria
#> 97             Planctomycetes
#> 98             Proteobacteria
#> 99    Actinobacteria (phylum)
#> 100            Proteobacteria
#> 101            Proteobacteria
#> 102   Actinobacteria (phylum)
#> 103            Proteobacteria
#> 104             Bacteroidetes
#> 105            Proteobacteria
#> 106            Proteobacteria
#> 107   Actinobacteria (phylum)
#> 108            Proteobacteria
#> 109    Acidobacteria (phylum)
#> 110   Actinobacteria (phylum)
#> 111            Proteobacteria
#> 112   Actinobacteria (phylum)
#> 113   Actinobacteria (phylum)
#> 114            Proteobacteria
#> 115            Proteobacteria
#> 116           Armatimonadetes
#> 117           Verrucomicrobia
#> 118                Firmicutes
#> 119   Actinobacteria (phylum)
#> 120           Verrucomicrobia
#> 121            Proteobacteria
#> 122 Gemmatimonadetes (phylum)
#> 123   Actinobacteria (phylum)
#> 124            Proteobacteria
#> 125             Cyanobacteria
#> 126           Verrucomicrobia
#> 127   Actinobacteria (phylum)
#> 128                Firmicutes
#> 129            Proteobacteria
#> 130            Planctomycetes
#> 131            Proteobacteria
#> 132            Proteobacteria
#> 133            Proteobacteria
#> 134                Firmicutes
#> 135   Actinobacteria (phylum)
#> 136   Actinobacteria (phylum)
#> 137                Firmicutes
#> 138                    Thermi
#> 139     Spirochaetes (phylum)
#> 140                Firmicutes
#> 141            Proteobacteria
#> 142            Proteobacteria
#> 143                Firmicutes
#> 144             Lentisphaerae
#> 145                       TM7
#> 146             Bacteroidetes
#> 147            Proteobacteria
#> 148                Firmicutes
#> 149   Actinobacteria (phylum)
#> 150             Euryarchaeota
#> 151   Actinobacteria (phylum)
#> 152             Cyanobacteria