This function identifies the most abundant taxa and distinguishes between nodes with statistically significant differences in their abundance (red) and those without (blue).

This function identifies the most abundant taxa and distinguishes between nodes with statistically significant differences in their abundance (red) and those without (blue).

Usage

diff_nodes_graph(T_Collapsed, n, mat_list, g.list, alpha)

Arguments

n

Number of nodes to consider.

mat_list

List of abundance matrices (preprocessed by T_collapse).

g.list

List of igraph objects (multilayer network), just 2 layers.

alpha

Significance. It must be a number between 0 and 1.

T_collapsed

Abundance table collapsed at a certain taxonomic level (preprocessed by T_collapse).

Value

Returns a plot with both networks collapsed.

Examples

data ("tc_dstoolA")
data ("mlnet_dstoolAB")

mlst <- list(tc_dstoolA, tc_dstoolB)

diff_nodes_graph (tc_dstoolA, n=10, mat_list=mlst, g.list=mlnet_dstoolAB, alpha=0.5)
#> Loading required package: muxViz
#> IGRAPH 3f47a03 UNW- 152 22 -- 
#> + attr: name (v/c), colorA (v/c), weight (e/n)
#> + edges from 3f47a03 (vertex names):
#>  [1] Acidobacteriaceae --Sinobacteraceae   
#>  [2] Oceanospirillaceae--Bacteriovoracaceae
#>  [3] Oceanospirillaceae--Methylophilaceae  
#>  [4] Bacteriovoracaceae--Methylophilaceae  
#>  [5] Oceanospirillaceae--Methylocystaceae  
#>  [6] Bacteriovoracaceae--Methylocystaceae  
#>  [7] Methylophilaceae  --Methylocystaceae  
#>  [8] Acidobacteriaceae --Streptomycetaceae 
#> + ... omitted several edges