Generating a data.frame of node degrees per layer.

Assigns the taxonomic membership of the nodes.

Usage

ml_TaxGroup (g.list, T_table, g_tax, p_tax)

Arguments

g.list

List of igraph objects (multilayer network).

T_table

Taxa table.

g_tax

The taxonomic level at which the nodes will be assigned.

p_tax

The taxonomic level at which the network was built.

Value

The function returns a g.list (multilayer network) with each node named according its taxonomic level.

Examples

library (seqtime)
# Load g.list
data ("mlnet_dstoolAB")

# Replace columns' names of T_table
colnames (david_stool_lineages) <- c("ID", "Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species", "Summary", "ID")

ml_TaxGroup(mlnet_dstoolAB, T_table=david_stool_lineages, g_tax="Phylum", p_tax="Family")
#> [[1]]
#> IGRAPH 2d45ded UN-- 152 4 -- 
#> + attr: name (v/c), Taxon (v/x)
#> + edges from 2d45ded (vertex names):
#> [1] Tsukamurellaceae   --Gallionellaceae                              
#> [2] Hyphomonadaceae    --Armatimonadaceae                             
#> [3] Alicyclobacillaceae--ACK-M1                                       
#> [4] Nitrospiraceae     --ThermoanaerobacteralesFamilyIII.IncertaeSedis
#> 
#> [[2]]
#> IGRAPH 6e2239b UN-- 152 17 -- 
#> + attr: name (v/c), Taxon (v/x)
#> + edges from 6e2239b (vertex names):
#>  [1] Acidobacteriaceae --Sinobacteraceae   
#>  [2] Oceanospirillaceae--Bacteriovoracaceae
#>  [3] Oceanospirillaceae--Methylophilaceae  
#>  [4] Bacteriovoracaceae--Methylophilaceae  
#>  [5] Oceanospirillaceae--Methylocystaceae  
#>  [6] Bacteriovoracaceae--Methylocystaceae  
#>  [7] Methylophilaceae  --Methylocystaceae  
#>  [8] Acidobacteriaceae --Streptomycetaceae 
#> + ... omitted several edges
#>