This function displays the centrality measure of each phylum.

This function displays the centrality measure of each phylum.

Usage

phyl_ctr_plot (phyl_ctr_df, layer_name, title)

Arguments

phyl_ctr_df

A data.frame preprocessed by phyl_ctr_df

layer_name

Name of the layer to graph/plot.

title

Name for the graph.

Value

A bar plot of degree importance for each phylum.

Examples

data ("phyl_ctr_dstool")

n <- c("Sample A", "Sample B")

phyl_ctr_plot (phyl_ctr_dstool, layer_name = "Sample A", title = "Graph 1")
#> Don't know how to automatically pick scale for object of type <data.frame>.
#> Defaulting to continuous.
#> Warning: Use of `phyl_ctr_df$Taxon` is discouraged.
#> ℹ Use `Taxon` instead.
#> Error in geom_bar(stat = "identity"): Problem while computing aesthetics.
#> ℹ Error occurred in the 1st layer.
#> Caused by error in `check_aesthetics()`:
#> ! Aesthetics must be either length 1 or the same as the data (8).
#> ✖ Fix the following mappings: `y`.