This function aggregates the abundance data at a specified taxonomic level. It accepts a phyloseq object or a data.frame as input.

This function aggregates the abundance data at a specified taxonomic level. It accepts a phyloseq object or a data.frame as input.

Usage

T_collapse (is_phyloseq=T, ps=..., names_level=...)

Arguments

ps

If "is_phyloseq = T", a phyloseq object, and "is_phyloseq = F" for data.frame.

T_table

(If is_phyloseq = F) A taxa table.

O_table

(If is_phyloseq = F) An OTUs/ASV table.

names_level

(If is_phyloseq = F) Taxonomic level to group (ex. class, genus).

Value

Returns a class array matrix at a selected taxonomic level.

Examples

# With BiocManager, package "microbiome"
library(microbiome)
#> Loading required package: phyloseq
#> Loading required package: ggplot2
#> Want to understand how all the pieces fit together? Read R for Data
#> Science: https://r4ds.hadley.nz/
#> 
#> microbiome R package (microbiome.github.com)
#>     
#> 
#> 
#>  Copyright (C) 2011-2022 Leo Lahti, 
#>     Sudarshan Shetty et al. <microbiome.github.io>
#> 
#> Attaching package: ‘microbiome’
#> The following object is masked from ‘package:ggplot2’:
#> 
#>     alpha
#> The following object is masked from ‘package:base’:
#> 
#>     transform
# if is_phyloseq = T

library (microbiome)
data ("dietswap")
T_collapse (is_phyloseq = T, ps = dietswap, names_level ="Phylum")
#>                                       Actinobacteria Firmicutes Proteobacteria
#> Actinomycetaceae                                   1          5              2
#> Aerococcus                                         0          0              0
#> Aeromonas                                          0          2              0
#> Akkermansia                                      361       7112           4514
#> Alcaligenes faecalis et rel.                      21        378            102
#> Allistipes et rel.                               797      15042           5543
#> Anaerobiospirillum                                 0          0              1
#> Anaerofustis                                       0          2              0
#> Anaerostipes caccae et rel.                      667       4296           1415
#> Anaerotruncus colihominis et rel.                496       5276           1642
#> Anaerovorax odorimutans et rel.                  126       1049            371
#> Aneurinibacillus                                   0          0              0
#> Aquabacterium                                      5         20             20
#> Asteroleplasma et rel.                             0          0              0
#> Atopobium                                          2         10              4
#> Bacillus                                           9         79             27
#> Bacteroides fragilis et rel.                    1958      12241           2490
#> Bacteroides intestinalis et rel.                  71        809            316
#> Bacteroides ovatus et rel.                       825       7074           1899
#> Bacteroides plebeius et rel.                     237       3096            925
#> Bacteroides splachnicus et rel.                  781       5706            972
#> Bacteroides stercoris et rel.                    182       2347            509
#> Bacteroides uniformis et rel.                    720       5034           2413
#> Bacteroides vulgatus et rel.                    6239      75103          24748
#> Bifidobacterium                                  318       4115           2167
#> Bilophila et rel.                                 12        118             38
#> Brachyspira                                        3         10              5
#> Bryantella formatexigens et rel.                 411       5847           2944
#> Bulleidia moorei et rel.                          35        355            299
#> Burkholderia                                      13         87             25
#> Butyrivibrio crossotus et rel.                  1288      11585           5382
#> Campylobacter                                     33        305            105
#> Catenibacterium mitsuokai et rel.                 28        373            291
#> Clostridium (sensu stricto)                      251       1826            750
#> Clostridium cellulosi et rel.                   6033      64772          20866
#> Clostridium colinum et rel.                      316       2003            755
#> Clostridium difficile et rel.                    442       1982            732
#> Clostridium felsineum et rel.                      0          0              0
#> Clostridium leptum et rel.                       430       6863           2231
#> Clostridium nexile et rel.                       689       3416           1251
#> Clostridium orbiscindens et rel.                2709      17400           5599
#> Clostridium ramosum et rel.                       20        187             64
#> Clostridium sphenoides et rel.                   644       4830           2282
#> Clostridium stercorarium et rel.                  63        611            120
#> Clostridium symbiosum et rel.                   2126      15586           6752
#> Clostridium thermocellum et rel.                   0          0              0
#> Collinsella                                      132        727            492
#> Coprobacillus catenaformis et rel.                30        275             97
#> Coprococcus eutactus et rel.                    1389       6724           2926
#> Corynebacterium                                    8         76             26
#> Desulfovibrio et rel.                             32        350             98
#> Dialister                                         73       9515           3679
#> Dorea formicigenerans et rel.                   1572       7146           3738
#> Eggerthella lenta et rel.                         37        318            131
#> Enterobacter aerogenes et rel.                   156       2719            305
#> Enterococcus                                      44       1312            110
#> Escherichia coli et rel.                         875      10487            682
#> Eubacterium biforme et rel.                      388       2970           1020
#> Eubacterium cylindroides et rel.                  17        194             72
#> Eubacterium hallii et rel.                       179       1667            626
#> Eubacterium limosum et rel.                        8         81             32
#> Eubacterium rectale et rel.                      609       3686           1161
#> Eubacterium siraeum et rel.                       23        242             85
#> Eubacterium ventriosum et rel.                   137       1403            495
#> Faecalibacterium prausnitzii et rel.            3844      32452          13818
#> Fusobacteria                                      40        553            251
#> Gemella                                            0          1              3
#> Granulicatella                                     0         27              1
#> Haemophilus                                       12        361             33
#> Helicobacter                                      17        161             53
#> Klebisiella pneumoniae et rel.                    81       1892            182
#> Lachnobacillus bovis et rel.                     883       3534           1106
#> Lachnospira pectinoschiza et rel.                767       3982           1190
#> Lactobacillus catenaformis et rel.                 4         60             66
#> Lactobacillus gasseri et rel.                     60        584            220
#> Lactobacillus plantarum et rel.                   94        947           1752
#> Lactobacillus salivarius et rel.                  19        175            134
#> Lactococcus                                       10        108             29
#> Leminorella                                        3         37             16
#> Megamonas hypermegale et rel.                      9         81             26
#> Megasphaera elsdenii et rel.                     198       4656           1183
#> Methylobacterium                                   0          0              0
#> Micrococcaceae                                     0          0              0
#> Mitsuokella multiacida et rel.                   813       5080           1229
#> Moraxellaceae                                      5         45             11
#> Novosphingobium                                    0          0              0
#> Oceanospirillum                                   11        127             36
#> Oscillospira guillermondii et rel.             17998     113445          38934
#> Outgrouping clostridium cluster XIVa             708       3452           1571
#> Oxalobacter formigenes et rel.                   195       2247            939
#> Papillibacter cinnamivorans et rel.              303       2150            924
#> Parabacteroides distasonis et rel.               541       7144           2371
#> Peptococcus niger et rel.                         34        291             84
#> Peptostreptococcus anaerobius et rel.              0          0              0
#> Peptostreptococcus micros et rel.                 17        165             55
#> Phascolarctobacterium faecium et rel.            292       3407            433
#> Prevotella melaninogenica et rel.              47519     335833         105258
#> Prevotella oralis et rel.                       7291      51478          16130
#> Prevotella ruminicola et rel.                     30        245             68
#> Prevotella tannerae et rel.                      248       4285            907
#> Propionibacterium                                  9         86             31
#> Proteus et rel.                                   38        406            118
#> Pseudomonas                                        1          3              1
#> Roseburia intestinalis et rel.                   192       1435            597
#> Ruminococcus bromii et rel.                      405       4534           1837
#> Ruminococcus callidus et rel.                    262       4902           1329
#> Ruminococcus gnavus et rel.                      293       1888            841
#> Ruminococcus lactaris et rel.                    252        764            314
#> Ruminococcus obeum et rel.                      2016      20308           8005
#> Serratia                                          27        754             49
#> Sporobacter termitidis et rel.                  1863      18101           6304
#> Staphylococcus                                     1          6              3
#> Streptococcus bovis et rel.                      311       6074           1379
#> Streptococcus intermedius et rel.                 25        527            153
#> Streptococcus mitis et rel.                      201       3573            951
#> Subdoligranulum variable at rel.                1594      13400           6629
#> Sutterella wadsworthia et rel.                   458       3160           1086
#> Tannerella et rel.                               203       2235            801
#> Uncultured Bacteroidetes                          12        153             51
#> Uncultured Chroococcales                           0          0              0
#> Uncultured Clostridiales I                       786       7328           2829
#> Uncultured Clostridiales II                      566       3742           2294
#> Uncultured Mollicutes                            504       5092            992
#> Uncultured Selenomonadaceae                        0        140             34
#> Veillonella                                       56        691            510
#>                                       Verrucomicrobia Bacteroidetes
#> Actinomycetaceae                                    1             2
#> Aerococcus                                          0             0
#> Aeromonas                                           0             0
#> Akkermansia                                       256          2959
#> Alcaligenes faecalis et rel.                        2           103
#> Allistipes et rel.                                 96          1282
#> Anaerobiospirillum                                  0             0
#> Anaerofustis                                        0             0
#> Anaerostipes caccae et rel.                        36           789
#> Anaerotruncus colihominis et rel.                  68          1727
#> Anaerovorax odorimutans et rel.                    60           425
#> Aneurinibacillus                                    0             0
#> Aquabacterium                                       0             2
#> Asteroleplasma et rel.                              0             0
#> Atopobium                                           1             7
#> Bacillus                                            1            16
#> Bacteroides fragilis et rel.                       29          2846
#> Bacteroides intestinalis et rel.                    6            70
#> Bacteroides ovatus et rel.                         46           466
#> Bacteroides plebeius et rel.                       15           283
#> Bacteroides splachnicus et rel.                    24           510
#> Bacteroides stercoris et rel.                      20           233
#> Bacteroides uniformis et rel.                       6          1118
#> Bacteroides vulgatus et rel.                      469          5083
#> Bifidobacterium                                   493          1973
#> Bilophila et rel.                                   1            19
#> Brachyspira                                         0             3
#> Bryantella formatexigens et rel.                   47           890
#> Bulleidia moorei et rel.                            3            71
#> Burkholderia                                        0            20
#> Butyrivibrio crossotus et rel.                    146          2865
#> Campylobacter                                       4            68
#> Catenibacterium mitsuokai et rel.                   2            62
#> Clostridium (sensu stricto)                       154           693
#> Clostridium cellulosi et rel.                     310         13213
#> Clostridium colinum et rel.                        10           491
#> Clostridium difficile et rel.                     546          1191
#> Clostridium felsineum et rel.                       0             0
#> Clostridium leptum et rel.                        214          1681
#> Clostridium nexile et rel.                         43          1129
#> Clostridium orbiscindens et rel.                  691          6835
#> Clostridium ramosum et rel.                         2            35
#> Clostridium sphenoides et rel.                     63          1275
#> Clostridium stercorarium et rel.                    8           129
#> Clostridium symbiosum et rel.                     111          3169
#> Clostridium thermocellum et rel.                    0             0
#> Collinsella                                        46           368
#> Coprobacillus catenaformis et rel.                  4            43
#> Coprococcus eutactus et rel.                      329          1904
#> Corynebacterium                                     1            16
#> Desulfovibrio et rel.                               8            72
#> Dialister                                          24          1663
#> Dorea formicigenerans et rel.                     135          2198
#> Eggerthella lenta et rel.                          18            92
#> Enterobacter aerogenes et rel.                     36           465
#> Enterococcus                                        5           109
#> Escherichia coli et rel.                           33           630
#> Eubacterium biforme et rel.                        65           487
#> Eubacterium cylindroides et rel.                    2            44
#> Eubacterium hallii et rel.                         31           393
#> Eubacterium limosum et rel.                         1            23
#> Eubacterium rectale et rel.                        15           885
#> Eubacterium siraeum et rel.                         3            61
#> Eubacterium ventriosum et rel.                     11           291
#> Faecalibacterium prausnitzii et rel.              365          9924
#> Fusobacteria                                        5            94
#> Gemella                                             0             0
#> Granulicatella                                      0             1
#> Haemophilus                                         0            47
#> Helicobacter                                        2            35
#> Klebisiella pneumoniae et rel.                     22           333
#> Lachnobacillus bovis et rel.                       26           706
#> Lachnospira pectinoschiza et rel.                  55          1027
#> Lactobacillus catenaformis et rel.                  0             6
#> Lactobacillus gasseri et rel.                       8           122
#> Lactobacillus plantarum et rel.                    14           206
#> Lactobacillus salivarius et rel.                   16            47
#> Lactococcus                                        81            23
#> Leminorella                                         0            10
#> Megamonas hypermegale et rel.                       1            16
#> Megasphaera elsdenii et rel.                       31           983
#> Methylobacterium                                    0             0
#> Micrococcaceae                                      0             0
#> Mitsuokella multiacida et rel.                      5          1270
#> Moraxellaceae                                       0             9
#> Novosphingobium                                     0             0
#> Oceanospirillum                                     2            26
#> Oscillospira guillermondii et rel.               4279         38538
#> Outgrouping clostridium cluster XIVa               39          1100
#> Oxalobacter formigenes et rel.                     62           688
#> Papillibacter cinnamivorans et rel.                45           800
#> Parabacteroides distasonis et rel.                 19           650
#> Peptococcus niger et rel.                          10            84
#> Peptostreptococcus anaerobius et rel.               0             1
#> Peptostreptococcus micros et rel.                   3            42
#> Phascolarctobacterium faecium et rel.               4           300
#> Prevotella melaninogenica et rel.                 107         78343
#> Prevotella oralis et rel.                          10         11607
#> Prevotella ruminicola et rel.                       2            56
#> Prevotella tannerae et rel.                        15           480
#> Propionibacterium                                   1            18
#> Proteus et rel.                                     4            78
#> Pseudomonas                                         0             1
#> Roseburia intestinalis et rel.                      7           302
#> Ruminococcus bromii et rel.                        60          1393
#> Ruminococcus callidus et rel.                      38          1354
#> Ruminococcus gnavus et rel.                        41           553
#> Ruminococcus lactaris et rel.                      39           135
#> Ruminococcus obeum et rel.                        476          4986
#> Serratia                                            3            84
#> Sporobacter termitidis et rel.                    764          7421
#> Staphylococcus                                      0             0
#> Streptococcus bovis et rel.                        27          1163
#> Streptococcus intermedius et rel.                   9            83
#> Streptococcus mitis et rel.                        21           618
#> Subdoligranulum variable at rel.                  963          3912
#> Sutterella wadsworthia et rel.                      7           414
#> Tannerella et rel.                                 15           343
#> Uncultured Bacteroidetes                            5            30
#> Uncultured Chroococcales                            0             4
#> Uncultured Clostridiales I                        399          2525
#> Uncultured Clostridiales II                       155          2611
#> Uncultured Mollicutes                             205          1129
#> Uncultured Selenomonadaceae                         0            16
#> Veillonella                                         5            81
#>                                       Spirochaetes Fusobacteria Cyanobacteria
#> Actinomycetaceae                                 0            0             0
#> Aerococcus                                       0            0             0
#> Aeromonas                                        0            0             0
#> Akkermansia                                     36           51            65
#> Alcaligenes faecalis et rel.                     6            2             2
#> Allistipes et rel.                              74           17           219
#> Anaerobiospirillum                               0            0             0
#> Anaerofustis                                     0            0             0
#> Anaerostipes caccae et rel.                     14           12            23
#> Anaerotruncus colihominis et rel.              499           21            12
#> Anaerovorax odorimutans et rel.                 14            8            11
#> Aneurinibacillus                                 0            0             0
#> Aquabacterium                                    0            0             0
#> Asteroleplasma et rel.                           0            0             0
#> Atopobium                                        0            0             0
#> Bacillus                                         1            1             1
#> Bacteroides fragilis et rel.                   339           12            85
#> Bacteroides intestinalis et rel.                 3            1             3
#> Bacteroides ovatus et rel.                      27           11            44
#> Bacteroides plebeius et rel.                    17            8            28
#> Bacteroides splachnicus et rel.                 29           13            30
#> Bacteroides stercoris et rel.                   13            5            46
#> Bacteroides uniformis et rel.                  102            1            49
#> Bacteroides vulgatus et rel.                   412           44          1179
#> Bifidobacterium                                 42           16            61
#> Bilophila et rel.                                1            1             5
#> Brachyspira                                      1            0             0
#> Bryantella formatexigens et rel.                26           27            43
#> Bulleidia moorei et rel.                         2            4             3
#> Burkholderia                                     1            0             3
#> Butyrivibrio crossotus et rel.                 181           93            74
#> Campylobacter                                    4            4             4
#> Catenibacterium mitsuokai et rel.                1            4            20
#> Clostridium (sensu stricto)                    120           20            15
#> Clostridium cellulosi et rel.                 1569          335          3021
#> Clostridium colinum et rel.                      8            8             5
#> Clostridium difficile et rel.                   30           20             9
#> Clostridium felsineum et rel.                    0            0             0
#> Clostridium leptum et rel.                     116           15            91
#> Clostridium nexile et rel.                      11           70            14
#> Clostridium orbiscindens et rel.               321          210           192
#> Clostridium ramosum et rel.                      2            2             2
#> Clostridium sphenoides et rel.                  30           96            37
#> Clostridium stercorarium et rel.                 3            2             3
#> Clostridium symbiosum et rel.                   79          365           170
#> Clostridium thermocellum et rel.                 0            0             0
#> Collinsella                                      6           13             8
#> Coprobacillus catenaformis et rel.               2            2             3
#> Coprococcus eutactus et rel.                    25           96            74
#> Corynebacterium                                  1            1             1
#> Desulfovibrio et rel.                            3            4            33
#> Dialister                                       13           15            14
#> Dorea formicigenerans et rel.                   35          177           120
#> Eggerthella lenta et rel.                        4            4             4
#> Enterobacter aerogenes et rel.                   7           12            26
#> Enterococcus                                     4          137             4
#> Escherichia coli et rel.                         7          997          2036
#> Eubacterium biforme et rel.                      5          292            58
#> Eubacterium cylindroides et rel.                 2            2             5
#> Eubacterium hallii et rel.                       5           24            19
#> Eubacterium limosum et rel.                      1            1             1
#> Eubacterium rectale et rel.                     11           62            12
#> Eubacterium siraeum et rel.                      4            2             2
#> Eubacterium ventriosum et rel.                   5           14             8
#> Faecalibacterium prausnitzii et rel.           409          269            69
#> Fusobacteria                                     5            5             5
#> Gemella                                          0            0             0
#> Granulicatella                                   0            1             0
#> Haemophilus                                      0            7             2
#> Helicobacter                                     3            2             2
#> Klebisiella pneumoniae et rel.                   4           19            16
#> Lachnobacillus bovis et rel.                    15           33            15
#> Lachnospira pectinoschiza et rel.               20           71            14
#> Lactobacillus catenaformis et rel.               0            0             0
#> Lactobacillus gasseri et rel.                    7            6             7
#> Lactobacillus plantarum et rel.                 10           10            10
#> Lactobacillus salivarius et rel.                 1            2             1
#> Lactococcus                                      1            2             1
#> Leminorella                                      0            0             0
#> Megamonas hypermegale et rel.                    1            1             2
#> Megasphaera elsdenii et rel.                   128           62            49
#> Methylobacterium                                 0            0             0
#> Micrococcaceae                                   0            0             0
#> Mitsuokella multiacida et rel.                  97           16             2
#> Moraxellaceae                                    0            0             0
#> Novosphingobium                                  0            0             0
#> Oceanospirillum                                  1            1             1
#> Oscillospira guillermondii et rel.            4955          339           163
#> Outgrouping clostridium cluster XIVa            21          113            19
#> Oxalobacter formigenes et rel.                  51            7            15
#> Papillibacter cinnamivorans et rel.             14           25            23
#> Parabacteroides distasonis et rel.             143           10           125
#> Peptococcus niger et rel.                        8            3             2
#> Peptostreptococcus anaerobius et rel.            0            0             0
#> Peptostreptococcus micros et rel.                2            2             2
#> Phascolarctobacterium faecium et rel.           26           11           491
#> Prevotella melaninogenica et rel.             2589         3234            37
#> Prevotella oralis et rel.                      492          559            40
#> Prevotella ruminicola et rel.                    3            2             2
#> Prevotella tannerae et rel.                     30            6            75
#> Propionibacterium                                1            1             1
#> Proteus et rel.                                  4            5             6
#> Pseudomonas                                      0            0             0
#> Roseburia intestinalis et rel.                   7           16            12
#> Ruminococcus bromii et rel.                     41           24             5
#> Ruminococcus callidus et rel.                  391           13            13
#> Ruminococcus gnavus et rel.                     10           28            88
#> Ruminococcus lactaris et rel.                    5           10             3
#> Ruminococcus obeum et rel.                     201          200           718
#> Serratia                                         0           26             1
#> Sporobacter termitidis et rel.                1050           45            36
#> Staphylococcus                                   0            0             0
#> Streptococcus bovis et rel.                     17           18            13
#> Streptococcus intermedius et rel.                3            9             6
#> Streptococcus mitis et rel.                     39           63            46
#> Subdoligranulum variable at rel.                85          119           346
#> Sutterella wadsworthia et rel.                  43           22            98
#> Tannerella et rel.                              36            9            28
#> Uncultured Bacteroidetes                         1            0             1
#> Uncultured Chroococcales                         0            0             0
#> Uncultured Clostridiales I                      95           18            18
#> Uncultured Clostridiales II                    286           38            17
#> Uncultured Mollicutes                            9           10             8
#> Uncultured Selenomonadaceae                      0            0            11
#> Veillonella                                      3           16             2
#>  [ reached getOption("max.print") -- omitted 5 rows ]

# if is_phyloseq = F

library (seqtime)
#> Loading required package: igraph
#> 
#> Attaching package: ‘igraph’
#> The following object is masked from ‘package:microbiome’:
#> 
#>     diversity
#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum
#> The following object is masked from ‘package:base’:
#> 
#>     union
#> Loading required package: vegan
#> Loading required package: permute
#> 
#> Attaching package: ‘permute’
#> The following object is masked from ‘package:igraph’:
#> 
#>     permute
#> Loading required package: lattice
#> This is vegan 2.6-10
#> 
#> Attaching package: ‘vegan’
#> The following object is masked from ‘package:igraph’:
#> 
#>     diversity
#> The following object is masked from ‘package:microbiome’:
#> 
#>     diversity
#> Registered S3 method overwritten by 'untb':
#>   method       from 
#>   plot.preston vegan
#> 
#> seqtime: Time Series Analysis of Sequencing Data
#> 
#> Attaching package: ‘seqtime’
#> The following object is masked from ‘package:igraph’:
#> 
#>     normalize
colnames (david_stool_lineages)<- c("ID", "Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species", "Summary", "ID")

T_collapse(is_phyloseq = F, T_table = david_stool_lineages, O_table = david_stoolA_otus, names_level = "Family")
#>                 Nitrospiraceae  none Rikenellaceae Porphyromonadaceae
#> SAMPLE_Stool1                0  5750           118                859
#> SAMPLE_Stool2                0  7719           125                299
#> SAMPLE_Stool3                0  7336            93                145
#> SAMPLE_Stool4                0   405             0                  0
#> SAMPLE_Stool5                0  7180            38                120
#> SAMPLE_Stool6                0 10115           727                658
#>                 Bacteroidaceae Prevotellaceae Flavobacteriaceae
#> SAMPLE_Stool1            17775              4                 5
#> SAMPLE_Stool2            23449             15                 0
#> SAMPLE_Stool3            23991             15                 0
#> SAMPLE_Stool4                0              0                 0
#> SAMPLE_Stool5            25269             34                 0
#> SAMPLE_Stool6            40196              5                 0
#>                 Acidobacteriaceae Spirochaetaceae Hyphomicrobiaceae
#> SAMPLE_Stool1                   0               0                 0
#> SAMPLE_Stool2                   0               0                 0
#> SAMPLE_Stool3                   0               0                 0
#> SAMPLE_Stool4                   0               0                 0
#> SAMPLE_Stool5                   0               0                 0
#> SAMPLE_Stool6                   0               0                 0
#>                 Bradyrhizobiaceae Xanthobacteraceae Methylobacteriaceae
#> SAMPLE_Stool1                   0                 0                   0
#> SAMPLE_Stool2                   0                 0                   0
#> SAMPLE_Stool3                   0                 0                   0
#> SAMPLE_Stool4                  93                 0                   0
#> SAMPLE_Stool5                   0                 0                   0
#> SAMPLE_Stool6                  11                 0                  21
#>                 Rhizobiaceae Caulobacteraceae Rhodobacteraceae
#> SAMPLE_Stool1              0                0                0
#> SAMPLE_Stool2              0                0                0
#> SAMPLE_Stool3              0                0                0
#> SAMPLE_Stool4              0                0                0
#> SAMPLE_Stool5              0                0                0
#> SAMPLE_Stool6              0                0                0
#>                 Sphingomonadaceae Neisseriaceae Rhodocyclaceae Alcaligenaceae
#> SAMPLE_Stool1                   0             1              0            345
#> SAMPLE_Stool2                   0            26              0            463
#> SAMPLE_Stool3                   0             2              0            484
#> SAMPLE_Stool4                   0             0              0              0
#> SAMPLE_Stool5                   0             1              0             22
#> SAMPLE_Stool6                   0             2              0             53
#>                 Comamonadaceae Burkholderiaceae Oxalobacteraceae
#> SAMPLE_Stool1                8                0                0
#> SAMPLE_Stool2               21                0                0
#> SAMPLE_Stool3                1                1                0
#> SAMPLE_Stool4              462                0                0
#> SAMPLE_Stool5                9                0                0
#> SAMPLE_Stool6               67                0               58
#>                 Cardiobacteriaceae Oceanospirillaceae Moraxellaceae
#> SAMPLE_Stool1                    0                  0            21
#> SAMPLE_Stool2                    1                  0             6
#> SAMPLE_Stool3                    0                  0            28
#> SAMPLE_Stool4                    0                  0             0
#> SAMPLE_Stool5                    0                  0             1
#> SAMPLE_Stool6                    0                  0             8
#>                 Pseudomonadaceae Pseudoalteromonadaceae Shewanellaceae
#> SAMPLE_Stool1                  1                      0              0
#> SAMPLE_Stool2                  1                      0              0
#> SAMPLE_Stool3                  0                      0              0
#> SAMPLE_Stool4                  0                      0              0
#> SAMPLE_Stool5                  0                      0              0
#> SAMPLE_Stool6                  0                      0              0
#>                 Vibrionaceae Pasteurellaceae Enterobacteriaceae
#> SAMPLE_Stool1              0             170                 86
#> SAMPLE_Stool2              0             305                 28
#> SAMPLE_Stool3              0             298                 19
#> SAMPLE_Stool4              0               0                  0
#> SAMPLE_Stool5              0              26                 17
#> SAMPLE_Stool6              0              49                 86
#>                 Desulfovibrionaceae Campylobacteraceae Fusobacteriaceae
#> SAMPLE_Stool1                   133                  3                0
#> SAMPLE_Stool2                    45                  2                0
#> SAMPLE_Stool3                    24                  7                0
#> SAMPLE_Stool4                     0                  0                0
#> SAMPLE_Stool5                    20                  0                0
#> SAMPLE_Stool6                   161                  3                0
#>                 Coriobacteriaceae Geodermatophilaceae Frankiaceae
#> SAMPLE_Stool1                  25                   0           0
#> SAMPLE_Stool2                  77                   0           0
#> SAMPLE_Stool3                  32                   0           0
#> SAMPLE_Stool4                   0                   0           0
#> SAMPLE_Stool5                  35                   0           0
#> SAMPLE_Stool6                 132                   0           0
#>                 Microbacteriaceae Cellulomonadaceae Intrasporangiaceae
#> SAMPLE_Stool1                   0                 0                  0
#> SAMPLE_Stool2                   0                 0                  0
#> SAMPLE_Stool3                   1                 0                  0
#> SAMPLE_Stool4                   0                 0                  0
#> SAMPLE_Stool5                   0                 0                  0
#> SAMPLE_Stool6                   0                 0                  0
#>                 Micrococcaceae Dermabacteraceae Dermacoccaceae Actinomycetaceae
#> SAMPLE_Stool1                6                0              0               15
#> SAMPLE_Stool2                1                0              0               28
#> SAMPLE_Stool3                0                0              0               16
#> SAMPLE_Stool4                0                0              0                0
#> SAMPLE_Stool5                0                0              0               21
#> SAMPLE_Stool6                0                0              0               27
#>                 Bifidobacteriaceae Propionibacteriaceae Mycobacteriaceae
#> SAMPLE_Stool1                 2248                    0                0
#> SAMPLE_Stool2                 8649                    0                0
#> SAMPLE_Stool3                 7247                    0                0
#> SAMPLE_Stool4                    0                    0                0
#> SAMPLE_Stool5                 9077                    0                0
#> SAMPLE_Stool6                13254                    0                0
#>                 Gordoniaceae Corynebacteriaceae Nocardiaceae Veillonellaceae
#> SAMPLE_Stool1              0                  1            0            2150
#> SAMPLE_Stool2              0                  2            0            1564
#> SAMPLE_Stool3              0                 18            0            1395
#> SAMPLE_Stool4              0                211            0               0
#> SAMPLE_Stool5              0                  0            0            1411
#> SAMPLE_Stool6              0                  1            0            1770
#>                 Peptococcaceae Lachnospiraceae
#> SAMPLE_Stool1                0           24917
#> SAMPLE_Stool2                0           45793
#> SAMPLE_Stool3                0           40150
#> SAMPLE_Stool4                0               0
#> SAMPLE_Stool5                0           35777
#> SAMPLE_Stool6                0           48166
#>                 ClostridialesFamilyXIII.IncertaeSedis Clostridiaceae
#> SAMPLE_Stool1                                       3            104
#> SAMPLE_Stool2                                       1             66
#> SAMPLE_Stool3                                       0             40
#> SAMPLE_Stool4                                       0              0
#> SAMPLE_Stool5                                       5             82
#> SAMPLE_Stool6                                       9             32
#>                 Peptostreptococcaceae ClostridialesFamilyXI.IncertaeSedis
#> SAMPLE_Stool1                       0                                  13
#> SAMPLE_Stool2                       0                                   5
#> SAMPLE_Stool3                       0                                  13
#> SAMPLE_Stool4                       0                                   0
#> SAMPLE_Stool5                       0                                  16
#> SAMPLE_Stool6                       0                                  21
#>                 Paenibacillaceae Bacillaceae Planococcaceae Gemellaceae
#> SAMPLE_Stool1                  0           0              0           2
#> SAMPLE_Stool2                  0           0              0           8
#> SAMPLE_Stool3                  0           0              0           5
#> SAMPLE_Stool4                  0           0              0           0
#> SAMPLE_Stool5                  0           0              0           1
#> SAMPLE_Stool6                  0           0              0           8
#>                 Aerococcaceae Carnobacteriaceae Leuconostocaceae
#> SAMPLE_Stool1               0                12               49
#> SAMPLE_Stool2               1                21               32
#> SAMPLE_Stool3               0                14               31
#> SAMPLE_Stool4              44                 0                0
#> SAMPLE_Stool5               0                 8               30
#> SAMPLE_Stool6               0                10               39
#>                 Lactobacillaceae Enterococcaceae Streptococcaceae
#> SAMPLE_Stool1                  2               1              135
#> SAMPLE_Stool2                  2               3              544
#> SAMPLE_Stool3                  2               2              281
#> SAMPLE_Stool4                  0               0                1
#> SAMPLE_Stool5                 51               0              316
#> SAMPLE_Stool6                 93               1             3706
#>                 Erysipelotrichaceae Ruminococcaceae Staphylococcaceae
#> SAMPLE_Stool1                  2016           23385                 0
#> SAMPLE_Stool2                  4681           32174                 0
#> SAMPLE_Stool3                  3302           31288                 0
#> SAMPLE_Stool4                     0               1                34
#> SAMPLE_Stool5                   537           36646                 0
#> SAMPLE_Stool6                  2245           60176                 0
#>                 Coxiellaceae Nocardioidaceae Brucellaceae Flexibacteraceae
#> SAMPLE_Stool1              0               0            0                0
#> SAMPLE_Stool2              0               0            0                0
#> SAMPLE_Stool3              0               0            0                0
#> SAMPLE_Stool4              0               0            0                0
#> SAMPLE_Stool5              0               0            0                0
#> SAMPLE_Stool6              0               0            0                0
#>                 Turicibacteraceae Tsukamurellaceae Phyllobacteriaceae
#> SAMPLE_Stool1                   2                0                  0
#> SAMPLE_Stool2                   1                0                  0
#> SAMPLE_Stool3                  22                0                  0
#> SAMPLE_Stool4                   0                0                  0
#> SAMPLE_Stool5                   4                0                  0
#> SAMPLE_Stool6                   0                0                  0
#>                 Brevibacteriaceae Aeromonadaceae Sinobacteraceae
#> SAMPLE_Stool1                   0              0               0
#> SAMPLE_Stool2                   0              0               0
#> SAMPLE_Stool3                   0              0               0
#> SAMPLE_Stool4                   0              0               0
#> SAMPLE_Stool5                   0              0               0
#> SAMPLE_Stool6                   0              0               0
#>                 Beijerinckiaceae Catabacteriaceae Dietziaceae
#> SAMPLE_Stool1                  0                2           0
#> SAMPLE_Stool2                  0                3           0
#> SAMPLE_Stool3                  0                0           0
#> SAMPLE_Stool4                  0                0           0
#> SAMPLE_Stool5                  0                0           0
#> SAMPLE_Stool6                  0                5           0
#>                 Pseudonocardiaceae F16 Halomonadaceae Xanthomonadaceae
#> SAMPLE_Stool1                    0   0              0                2
#> SAMPLE_Stool2                    0   0              0                0
#> SAMPLE_Stool3                    0   0              0                0
#> SAMPLE_Stool4                    0   0              0              213
#> SAMPLE_Stool5                    0   0              0               34
#> SAMPLE_Stool6                    0   0              0                0
#>                 Sphingobacteriaceae Alteromonadaceae Dethiosulfovibrionaceae
#> SAMPLE_Stool1                     0                0                       0
#> SAMPLE_Stool2                     0                0                      13
#> SAMPLE_Stool3                     0                0                       0
#> SAMPLE_Stool4                     0                0                       0
#> SAMPLE_Stool5                     0                0                       0
#> SAMPLE_Stool6                     0                0                       2
#>                 Thermaceae Bacteriovoracaceae Acetobacteraceae Nostocaceae
#> SAMPLE_Stool1            0                  0                0           0
#> SAMPLE_Stool2            0                  0                0           0
#> SAMPLE_Stool3            0                  0                0           0
#> SAMPLE_Stool4            0                  0                0           0
#> SAMPLE_Stool5            0                  0                0           0
#> SAMPLE_Stool6            0                  0                0           0
#>                 Colwelliaceae Planctomycetaceae Hydrogenophilaceae
#> SAMPLE_Stool1               0                 0                  0
#> SAMPLE_Stool2               0                 0                  0
#> SAMPLE_Stool3               0                 0                  0
#> SAMPLE_Stool4               0                 0                  0
#> SAMPLE_Stool5               0                 0                  0
#> SAMPLE_Stool6               0                 0                  0
#>                 Nocardiopsaceae Rhodospirillaceae Methylophilaceae
#> SAMPLE_Stool1                 0                 0                0
#> SAMPLE_Stool2                 0                 0                0
#> SAMPLE_Stool3                 0                 0                0
#> SAMPLE_Stool4                 0                 0                0
#> SAMPLE_Stool5                 0                 0                0
#> SAMPLE_Stool6                 0                 0                0
#>                 Promicromonosporaceae OM60 Cryomorphaceae Methylocystaceae
#> SAMPLE_Stool1                       0    0              0                0
#> SAMPLE_Stool2                       0    0              0                0
#> SAMPLE_Stool3                       0    0              0                0
#> SAMPLE_Stool4                       0    0              0                0
#> SAMPLE_Stool5                       0    0              0                0
#> SAMPLE_Stool6                       0    0              0                0
#>                 Bdellovibrionaceae Streptomycetaceae Hyphomonadaceae
#> SAMPLE_Stool1                    0                 0               0
#> SAMPLE_Stool2                    0                 0               0
#> SAMPLE_Stool3                    0                 0               0
#> SAMPLE_Stool4                    0                 0               0
#> SAMPLE_Stool5                    0                 0               0
#> SAMPLE_Stool6                    0                 0               0
#>                 Solibacteraceae Rubrobacteraceae Chromatiaceae Bogoriellaceae
#> SAMPLE_Stool1                 0                0             0              0
#> SAMPLE_Stool2                 0                0             0              0
#> SAMPLE_Stool3                 0                0             0              0
#> SAMPLE_Stool4                 0                0             0              0
#> SAMPLE_Stool5                 0                0             0              0
#> SAMPLE_Stool6                 0                0             0              0
#>                 Williamsiaceae Legionellaceae Gallionellaceae Armatimonadaceae
#> SAMPLE_Stool1                0              0               0                0
#> SAMPLE_Stool2                0              0               0                0
#> SAMPLE_Stool3                0              0               0                0
#> SAMPLE_Stool4                0              0               0                0
#> SAMPLE_Stool5                0              0               0                0
#> SAMPLE_Stool6                0              0               0                0
#>                 Opitutaceae Alicyclobacillaceae Solirubrobacteraceae
#> SAMPLE_Stool1             0                   0                    0
#> SAMPLE_Stool2             0                   0                    0
#> SAMPLE_Stool3             0                   0                    0
#> SAMPLE_Stool4             0                   0                    0
#> SAMPLE_Stool5             0                   0                    0
#> SAMPLE_Stool6             0                   0                    0
#>                 Spartobacteriaceae Aurantimonadaceae Gemmatimonadaceae
#> SAMPLE_Stool1                    0                 0                 0
#> SAMPLE_Stool2                    0                 0                 0
#> SAMPLE_Stool3                    0                 0                 0
#> SAMPLE_Stool4                    0                 0                 0
#> SAMPLE_Stool5                    0                 0                 0
#> SAMPLE_Stool6                    0                 0                 0
#>                 Micromonosporaceae Saccharospirillaceae Xenococcaceae
#> SAMPLE_Stool1                    0                    0             0
#> SAMPLE_Stool2                    0                    0             0
#> SAMPLE_Stool3                    0                    0             0
#> SAMPLE_Stool4                    0                    0             0
#> SAMPLE_Stool5                    0                    0             0
#> SAMPLE_Stool6                    0                    0             0
#>                 Verrucomicrobiaceae Dermatophilaceae Listeriaceae
#> SAMPLE_Stool1                     0                0            0
#> SAMPLE_Stool2                     0                0            0
#> SAMPLE_Stool3                     0                0            0
#> SAMPLE_Stool4                     0                0            0
#> SAMPLE_Stool5                     0                0            0
#> SAMPLE_Stool6                     0                0            0
#>                 Psychromonadaceae Gemmataceae Syntrophobacteraceae
#> SAMPLE_Stool1                   0           0                    0
#> SAMPLE_Stool2                   0           0                    0
#> SAMPLE_Stool3                   0           0                    0
#> SAMPLE_Stool4                   0           0                    0
#> SAMPLE_Stool5                   0           0                    0
#> SAMPLE_Stool6                   0           0                    0
#>                 Haliangiaceae Idiomarinaceae Dehalobacteriaceae AKIW874 ACK-M1
#> SAMPLE_Stool1               0              0                  0       0      0
#> SAMPLE_Stool2               0              0                  0       0      0
#> SAMPLE_Stool3               0              0                  0       0      0
#> SAMPLE_Stool4               0              0                  0       0      0
#> SAMPLE_Stool5               0              0                  0       0      0
#> SAMPLE_Stool6               0              0                  0       0      0
#>                 Eubacteriaceae Deinococcaceae Leptospiraceae
#> SAMPLE_Stool1                3              0              0
#> SAMPLE_Stool2                4              0              0
#> SAMPLE_Stool3                0              0              0
#> SAMPLE_Stool4                0              0              0
#> SAMPLE_Stool5                1              0              0
#> SAMPLE_Stool6                4              0              0
#>                 ThermoanaerobacteralesFamilyIII.IncertaeSedis Nitrosomonadaceae
#> SAMPLE_Stool1                                               0                 0
#> SAMPLE_Stool2                                               0                 0
#> SAMPLE_Stool3                                               0                 0
#> SAMPLE_Stool4                                               0                 0
#> SAMPLE_Stool5                                               0                 0
#> SAMPLE_Stool6                                               0                 0
#>                 Erythrobacteraceae Thermicanaceae Victivallaceae A29
#> SAMPLE_Stool1                    0              0              0   0
#> SAMPLE_Stool2                    0              0              0   0
#> SAMPLE_Stool3                    0              0              0   0
#> SAMPLE_Stool4                    0              0              0   0
#> SAMPLE_Stool5                    0              0              0   0
#> SAMPLE_Stool6                    0              0              0   0
#>                 Rhodothermaceae Helicobacteraceae Exiguobacteraceae
#> SAMPLE_Stool1                 0                 0                 0
#> SAMPLE_Stool2                 0                 0                 0
#> SAMPLE_Stool3                 0                 0                 0
#> SAMPLE_Stool4                 0                 0                 0
#> SAMPLE_Stool5                 0                 0                 0
#> SAMPLE_Stool6                 0                 0                 0
#>                 Patulibacteraceae Methanobacteriaceae Sanguibacteraceae
#> SAMPLE_Stool1                   0                   0                 0
#> SAMPLE_Stool2                   0                   0                 0
#> SAMPLE_Stool3                   0                   0                 0
#> SAMPLE_Stool4                   0                   0                 0
#> SAMPLE_Stool5                   0                   0                 0
#> SAMPLE_Stool6                   0                   0                 0
#>                 Phormidiaceae
#> SAMPLE_Stool1               0
#> SAMPLE_Stool2               0
#> SAMPLE_Stool3               0
#> SAMPLE_Stool4               0
#> SAMPLE_Stool5               0
#> SAMPLE_Stool6               0
#>  [ reached getOption("max.print") -- omitted 323 rows ]