This function assigns colors to the nodes based on their abundance at the specified taxonomic level.

This function assigns colors to the nodes based on their abundance at the specified taxonomic level.

abs_mat (abs.list, g.list, n)

Arguments

abs.list

Abundance list, the same lists you used to generate the network.

g.list

List of igraph objects (multilayer network).

n

Vertex number. Notice that if n>20 the nodes won't show so big.

Value

Returns a matrix, each column is a layer, and the values of the rows corresponds to the size of each node.

Examples

g <- readRDS (file = "examples/Dstool_mlnet.rds")
#> Warning: cannot open compressed file 'examples/Dstool_mlnet.rds', probable reason 'No such file or directory'
#> Error in gzfile(file, "rb"): cannot open the connection

abs_mat (david_stool_lineages, g, 10)
#> Error in abs_mat(david_stool_lineages, g, 10): could not find function "abs_mat"